+++ bootstrap/app.php
@@ -4,6 +4,7 @@ use App\Console\Commands\ImportDemoMedia;
 use App\Console\Commands\ProjectInitCommand;
 use App\Console\Commands\ResetPrimitiveBlocksCommand;
 use App\Console\Commands\SearchRebuildCommand;
+use App\Console\Commands\SyncWebBlocksUiIconsCommand;
 use App\Console\Commands\SiteExportCommand;
 use App\Console\Commands\SiteImportCommand;
 use App\Console\Commands\SiteCloneCommand;
@@ -55,6 +56,7 @@ return Application::configure(basePath: dirname(__DIR__))
         ProjectInitCommand::class,
         ResetPrimitiveBlocksCommand::class,
         SearchRebuildCommand::class,
+        SyncWebBlocksUiIconsCommand::class,
         SiteExportCommand::class,
         SiteImportCommand::class,
         SiteCloneCommand::class,